moved config.tsx to src
